sls_monitor/requirements.txt
# 基础依赖
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0

# 设备通信
pyserial>=3.5.0

# GUI界面 (内置tkinter，无需额外安装)
# tkinter  # Python内置

# 科学计算和数据处理
scipy>=1.7.0
pandas>=1.3.0

# 可选依赖（根据功能需要）
matplotlib>=3.4.0  # 用于数据可视化和绘图
scikit-image>=0.18.0  # 用于图像处理

# 开发和调试工具
argparse  # Python内置，命令行参数解析
datetime  # Python内置，时间处理
logging  # Python内置，日志记录
threading  # Python内置，多线程支持
pathlib  # Python内置，路径操作
struct  # Python内置，二进制数据处理
binascii  # Python内置，二进制/ASCII转换
itertools  # Python内置，迭代工具
math  # Python内置，数学函数
statistics  # Python内置，统计函数
random  # Python内置，随机数生成
time  # Python内置，时间相关
os  # Python内置，操作系统接口
sys  # Python内置，系统相关
typing  # Python内置，类型提示

# 注意：以上标注为"Python内置"的模块无需通过pip安装